requirements.txt
# PyTorch with CUDA support
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio

# ML & DL
huggingface_hub
torchtext
torchinfo
diffusers
wandb # ==0.13.10
accelerate
ftfy
safetensors
openai
sentence-transformers

# Data manipulation and analysis
einops
eindex-callum
pandas
numpy-stl
scikit-learn
hdbscan
umap-learn

# Utilities
requests
pytest
tqdm
ipykernel
notebook
ipywidgets==7.7.1
jupyterlab-widgets==1.1.1
jupyter
pre-commit
gdown
portalocker>=2.0.0
typeguard
jaxtyping
tiktoken
frozendict
tabulate
opencv-python
nbformat
ijson
googletrans

# Visualization
matplotlib
plotly
streamlit
streamlit-antd-components==0.2.5
streamlit-image-select
streamlit-on-Hover-tabs
nnsight>=0.3.5
imageio-ffmpeg
mediapy
seaborn


transformer_lens==2.11.0
sae-lens>=4.0.0,<5.0.0
sae-vis>=0.3.1
git+https://github.com/callummcdougall/CircuitsVis.git#subdirectory=python
git+https://github.com/neelnanda-io/neel-plotly

